+++ sdk/core/ts-http-runtime/src/util/aborterUtils.ts
@@ -31,7 +31,7 @@ export async function cancelablePromiseRace<T extends unknown[]>(
 ): Promise<T[number]> {
   const aborter = new AbortController();
   function abortHandler(): void {
-    aborter.abort();
+    aborter.abort(options?.abortSignal?.reason);
   }
   options?.abortSignal?.addEventListener("abort", abortHandler);
   try {